util/requirements.txt
absl-py==1.0.0
argon2-cffi==21.1.0
async-generator @ file:///home/ktietz/src/ci/async_generator_1611927993394/work
attrs @ file:///tmp/build/80754af9/attrs_1620827162558/work
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
bleach==4.1.0
certifi==2021.10.8
cffi @ file:///C:/ci_310/cffi_1642682485096/work
charset-normalizer==2.0.9
click==8.0.3
cog==0.0.3
colorama @ file:///tmp/build/80754af9/colorama_1607707115595/work
commonmark==0.9.1
contextlib2==21.6.0
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
Cython @ file:///C:/ci/cython_1639474807736/work
debugpy @ file:///C:/ci/debugpy_1637073815078/work
decorator @ file:///tmp/build/80754af9/decorator_1632776554403/work
defusedxml @ file:///tmp/build/80754af9/defusedxml_1615228127516/work
DNN-printer==0.0.2
docker-pycreds==0.4.0
easydict==1.9
einops==0.3.2
entrypoints==0.3
fonttools==4.28.2
gitdb==4.0.9
GitPython==3.1.26
grad-cam==1.3.7
idna==3.3
imageio==2.13.0
importlib-metadata @ file:///C:/ci/importlib-metadata_1631916848487/work
importlib-resources==5.4.0
ipykernel==6.5.1
ipython==7.30.0
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
ipywidgets @ file:///tmp/build/80754af9/ipywidgets_1634143127070/work
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
jsonschema==4.2.1
jupyter==1.0.0
jupyter-client==7.1.0
jupyter-console @ file:///tmp/build/80754af9/jupyter_console_1616615302928/work
jupyter-core @ file:///C:/ci/jupyter_core_1636537213213/work
jupyterlab-pygments @ file:///tmp/build/80754af9/jupyterlab_pygments_1601490720602/work
jupyterlab-widgets==1.0.2
kiwisolver @ file:///C:/ci/kiwisolver_1644962567532/work
koila==0.1
linformer==0.2.1
MarkupSafe @ file:///C:/ci/markupsafe_1621528314575/work
matplotlib @ file:///C:/ci/matplotlib-suite_1645455836921/work
matplotlib-inline==0.1.3
mistune==0.8.4
mkl-fft==1.3.1
mkl-random @ file:///C:/ci/mkl_random_1626186184278/work
mkl-service==2.4.0
ml-collections==0.1.0
msgpack @ file:///C:/ci/msgpack-python_1612287368835/work
munkres==1.1.4
natsort==8.0.2
nbclient==0.5.9
nbconvert==6.3.0
nbformat @ file:///tmp/build/80754af9/nbformat_1617383369282/work
nest-asyncio @ file:///tmp/build/80754af9/nest-asyncio_1613680548246/work
networkx==2.6.3
notebook @ file:///C:/ci/notebook_1637143794296/work
numpy @ file:///C:/ci/numpy_and_numpy_base_1634106873763/work
olefile @ file:///Users/ktietz/demo/mc3/conda-bld/olefile_1629805411829/work
opencv-python==4.5.4.60
packaging @ file:///tmp/build/80754af9/packaging_1637314298585/work
pandas==1.3.4
pandocfilters==1.5.0
parso==0.8.3
pathtools==0.1.2
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==8.4.0
prometheus-client @ file:///tmp/build/80754af9/prometheus_client_1637050397234/work
promise==2.3
prompt-toolkit==3.0.23
protobuf==3.19.4
psutil==5.9.0
pycocotools==2.0.4
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
Pygments @ file:///tmp/build/80754af9/pygments_1629234116488/work
pynvml==11.0.0
pyparsing==3.0.6
pyrsistent @ file:///C:/ci/pyrsistent_1636111468851/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz==2021.3
PyWavelets==1.2.0
pywin32==302
pywinpty==1.1.6
PyYAML==6.0
pyzmq @ file:///C:/ci/pyzmq_1638435185959/work
qtconsole @ file:///tmp/build/80754af9/qtconsole_1632739723211/work
QtPy @ file:///tmp/build/80754af9/qtpy_1629397026935/work
requests==2.26.0
rich==10.15.2
scikit-image==0.18.3
scikit-learn==1.0.1
scipy==1.8.0
Send2Trash @ file:///tmp/build/80754af9/send2trash_1632406701022/work
sentry-sdk==1.5.5
setproctitle==1.2.2
setuptools-scm==6.3.2
shortuuid==1.0.8
sip==4.19.13
six @ file:///tmp/build/80754af9/six_1623709665295/work
sklearn==0.0
smmap==5.0.0
tensorboardX @ file:///tmp/build/80754af9/tensorboardx_1621440489103/work
termcolor==1.1.0
terminado==0.12.1
testpath @ file:///tmp/build/80754af9/testpath_1624638946665/work
threadpoolctl==3.0.0
tifffile==2021.11.2
timm==0.4.12
tomli==1.2.2
torch==1.8.0
torchaudio==0.8.0
torchsummary==1.5.1
torchvision==0.9.0
tornado @ file:///C:/ci/tornado_1606942392901/work
tqdm==4.62.3
traitlets @ file:///tmp/build/80754af9/traitlets_1636710298902/work
ttach==0.0.3
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1631814937681/work
urllib3==1.26.7
vision-transformer-pytorch==1.0.3
wandb==0.12.11
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
webencodings==0.5.1
widgetsnbextension==3.5.2
wincertstore==0.2
WMI==1.5.1
yaspin==2.1.0
zipp @ file:///tmp/build/80754af9/zipp_1633618647012/work